feat: change position data int -> float

+++ tron/game_config.go
@@ -11,8 +11,8 @@ type Config struct {
 
 // Point .
 type Point struct {
-	X int
-	Y int
+	X float64
+	Y float64
 }
 
 // Vector .
